feat: use design theme as default theme
